Pin version of num-traits to avoid semver-breaking incompatibility
The num-traits package, which is an indirect dependency via
num-rational, landed an update this month which broke compilation
on rustc versions below 1.8. I don't think this is considered a
semver-breaking change on their part, since their README suggests
the package is only supported on 1.8 and newer in the first place.

I think it's fine to move to requiring 1.8 as a minimum version,
but before doing that, I want to publish a patch-level version that
fixes the semver-incompatible change for any users of this package.

By pinning the exact version of num-traits, we can offer one last
patch for this minor version release line that should continue to
work. Then, we can bump our version number and remove the pin.