Set up linting and format checking in CI
For now, this has to use the nightly version of Cargo/rustfmt, because I
really want to preserve the multiple newlines between sections of code,
and that requires the `blank_lines_upper_bound` configuration option in
`.rustfmt.toml`. See https://github.com/rust-lang/rustfmt/issues/3381
for the bug tracking the stabilization of this configuration option.